replaced  verifiers/miners pages
