+++ src/index.tsx
@@ -1,37 +1,34 @@
-import React from 'react';
-import ReactDOM from 'react-dom';
-import './index.scss';
-import { config } from './config';
-import App from './App';
-import WalletProvider from './context/Wallet/WalletProvider'
-import GithubProvider from './context/Github/GithubProvider'
-import DataProvider from './context/Data/DataProvider'
-import { Wallet } from './context/Wallet/Index'
-import { Github } from './context/Github/Index'
-import { Router, Route, Switch } from 'react-router-dom'
+import React from "react";
+import ReactDOM from "react-dom";
+import "./index.scss";
+import { config } from "./config";
+import App from "./App";
+import WalletProvider from "./context/Wallet/WalletProvider";
+import GithubProvider from "./context/Github/GithubProvider";
+import DataProvider from "./context/Data/DataProvider";
+import { Wallet } from "./context/Wallet/Index";
+import { Github } from "./context/Github/Index";
+import { Router, Route, Switch } from "react-router-dom";
 // @ts-ignore
 import { GlobalNotification, GlobalModal } from "slate-react-system";
-import * as serviceWorker from './serviceWorker';
-import Preonboarding from './pages/Preonboarding';
-import Onboarding from './pages/Onboarding';
-import Landing from './pages/Landing';
-import history from './context/History';
-import LogExplorer from './pages/LogExplorer';
-import './fonts/SuisseIntl-Regular.woff'
-import ClientDetails from './pages/ClientDetails'
-import { CookiesProvider } from 'react-cookie';
+import * as serviceWorker from "./serviceWorker";
+import Preonboarding from "./pages/Preonboarding";
+import Onboarding from "./pages/Onboarding";
+import Landing from "./pages/Landing";
+import history from "./context/History";
+import LogExplorer from "./pages/LogExplorer";
+import "./fonts/SuisseIntl-Regular.woff";
+import ClientDetails from "./pages/ClientDetails";
+import { CookiesProvider } from "react-cookie";
 import * as Sentry from "@sentry/react";
 import { Integrations } from "@sentry/tracing";
-import Footer from './components/Footer';
-import LdnApplication from './pages/LdnApplication';
-import Verifiers2 from './pages/Verifiers2';
-import Miners2 from './pages/Miners2';
-
-
-
+import Footer from "./components/Footer";
+import LdnApplication from "./pages/LdnApplication";
+import Verifiers from "./pages/Verifiers";
+import Miners from "./pages/Miners";
 
 // redirect to domain if user access fleek url
-if (window.location.host.includes('fleek') && config.willRedirect) {
+if (window.location.host.includes("fleek") && config.willRedirect) {
   window.location.href = config.domain;
 }
 
@@ -45,36 +42,42 @@ Sentry.init({
   tracesSampleRate: 1.0,
 });
 
-
-
-
 ReactDOM.render(
-  
-
   <React.StrictMode>
     <CookiesProvider>
       <WalletProvider>
         <GithubProvider>
           <Wallet.Consumer>
-            {wallet => (
+            {(wallet) => (
               <Github.Consumer>
-                {github => (
+                {(github) => (
                   <DataProvider wallet={wallet} github={github}>
                     <Router history={history}>
                       <Switch>
-                        <Route exact path={'/'} component={Onboarding} ></Route>
-                        <Route path={'/app'} component={App} ></Route>
-                        <Route path={'/wallet'} component={Preonboarding} ></Route>
-                        <Route path={'/landing'} component={Landing} ></Route>
-                        <Route path={'/verifiers'} component={Verifiers2} ></Route>
-                        <Route path={'/ldn-application'} component={LdnApplication} ></Route>
-                        <Route path={'/miners'} component={Miners2} ></Route>
-                        <Route path={'/client'} component={ClientDetails} ></Route>
-                        <Route path={'/logs'} component={LogExplorer} ></Route>                        
-                      </Switch>  
-                     <Footer />  
+                        <Route exact path={"/"} component={Onboarding}></Route>
+                        <Route path={"/app"} component={App}></Route>
+                        <Route
+                          path={"/wallet"}
+                          component={Preonboarding}
+                        ></Route>
+                        <Route path={"/landing"} component={Landing}></Route>
+                        <Route
+                          path={"/verifiers"}
+                          component={Verifiers}
+                        ></Route>
+                        <Route
+                          path={"/ldn-application"}
+                          component={LdnApplication}
+                        ></Route>
+                        <Route path={"/miners"} component={Miners}></Route>
+                        <Route
+                          path={"/client"}
+                          component={ClientDetails}
+                        ></Route>
+                        <Route path={"/logs"} component={LogExplorer}></Route>
+                      </Switch>
+                      <Footer />
                     </Router>
-                
                     <GlobalNotification style={{ bottom: 0, right: 0 }} />
                     <GlobalModal style={{ maxWidth: "none" }} />
                   </DataProvider>
@@ -86,7 +89,7 @@ ReactDOM.render(
       </WalletProvider>
     </CookiesProvider>
   </React.StrictMode>,
-  document.getElementById('root')
+  document.getElementById("root")
 );
 
 // If you want your app to work offline and load faster, you can change
